Fixed merge conflicts, need to test to make sure it still builds in Windows and in Linux

Conflicts:
	.gitignore
	example/media_tester/src/app/media_tester_app.cpp
	example/media_viewer/src/app/mediaviewer_app.cpp
	example/media_viewer/vs2013/media_viewer.sln
	projects/essentials/src/ds/debug/automator/automator.cpp
	projects/essentials/src/ds/ui/drawing/drawing_canvas.cpp
	projects/video/gstreamer-1.0/src/gstreamer/gstreamer_wrapper.cpp
	src/ds/app/environment.cpp
	src/ds/data/data_buffer.cpp
	src/ds/data/data_buffer.h
	src/ds/network/single_udp_receiver.cpp
	src/ds/network/udp_connection.cpp
	src/ds/ui/sprite/border.cpp
	src/ds/ui/sprite/fbo/fbo.h
	src/ds/ui/sprite/sprite_engine.cpp
	src/ds/ui/sprite/sprite_engine.h

+++ projects/physics/box2d/src/ds/physics/body_builder.cpp
@@ -119,7 +119,7 @@ void BodyBuilderPolygon::createFixture(SpriteBody& body) const
 	fixtureDef.filter.categoryBits = mCategoryBits;
 	fixtureDef.filter.maskBits = mMaskBits;
 
- 	int32 count = mPoints.size();
+ 	int32 count = (int32)mPoints.size();
  	b2Vec2 * vertices = new b2Vec2[count];
  	int i = 0;
  	for (auto it = mPoints.begin(); it < mPoints.end(); ++it){
