+++ src/ds/data/data_buffer.cpp
@@ -6,105 +6,99 @@
 namespace ds {
 
 template <>
-void DataBuffer::add<std::string>( const std::string &s )
-{
-  unsigned size = s.size();
-  add(size);
-  mStream.write(s.c_str(), size);
+void DataBuffer::add<std::string>(const std::string &s){
+	size_t size = (unsigned)s.size();
+	add(size);
+	mStream.write(s.c_str(), size);
 }
 
-void DataBuffer::add(const char *cs)
-{
-  add<std::string>(cs);
+void DataBuffer::add(const char *cs){
+	add<std::string>(cs);
 }
 
 template <>
-void DataBuffer::add<std::wstring>( const std::wstring &ws )
-{
-  unsigned size = (ws.size())*sizeof(wchar_t);
-  add(size);
-  mStream.write((const char *)(ws.c_str()), size);
+void DataBuffer::add<std::wstring>(const std::wstring &ws){
+	size_t size = (ws.size())*sizeof(wchar_t);
+	add(size);
+	mStream.write((const char *)(ws.c_str()), size);
 }
 
-void DataBuffer::add(const wchar_t *cs)
-{
-  add<std::wstring>(cs);
+void DataBuffer::add(const wchar_t *cs){
+	add<std::wstring>(cs);
 }
 
-void DataBuffer::add( const char *b, unsigned size )
-{
-  add(size);
-  mStream.write(b, size);
+void DataBuffer::add(const char *b, unsigned size){
+	add(size);
+	mStream.write(b, size);
 }
 
-bool DataBuffer::read( char *b, unsigned size )
-{
-  unsigned wsize = read<unsigned>();
-  if (wsize != size) {
-    add(wsize);
-    return false;
-  }
+bool DataBuffer::read(char *b, size_t size){
+	size_t wsize = read<size_t>();
+	if(wsize != size) {
+		add(wsize);
+		return false;
+	}
 
-  unsigned currentPosition = mStream.getReadPosition();
+	size_t currentPosition = mStream.getReadPosition();
 
-  mStream.setReadPosition(ReadWriteBuffer::End);
-  unsigned length = mStream.getReadPosition();
-  mStream.setReadPosition(currentPosition);
+	mStream.setReadPosition(ReadWriteBuffer::End);
+	size_t length = mStream.getReadPosition();
+	mStream.setReadPosition(currentPosition);
 
-  if (size > (length - currentPosition))
-    return false;
+	if(size > (length - currentPosition))
+		return false;
 
 
-  mStream.read(b, size);
-  return true;
+	mStream.read(b, size);
+	return true;
 }
 
-DataBuffer::DataBuffer(unsigned initialStreamSize)
-  : mStream(initialStreamSize)
+DataBuffer::DataBuffer(size_t initialStreamSize)
+	: mStream(initialStreamSize)
 {
 
 }
 
 void DataBuffer::seekBegin()
 {
-  mStream.setReadPosition(ReadWriteBuffer::Begin);
-  mStream.setWritePosition(ReadWriteBuffer::Begin);
+	mStream.setReadPosition(ReadWriteBuffer::Begin);
+	mStream.setWritePosition(ReadWriteBuffer::Begin);
 }
 
-unsigned DataBuffer::size()
+size_t DataBuffer::size()
 {
-  unsigned currentPosition = mStream.getReadPosition();
+	size_t currentPosition = mStream.getReadPosition();
 
-  mStream.setReadPosition(ReadWriteBuffer::End);
-  unsigned length = mStream.getReadPosition();
-  mStream.setReadPosition(currentPosition);
+	mStream.setReadPosition(ReadWriteBuffer::End);
+	size_t length = mStream.getReadPosition();
+	mStream.setReadPosition(currentPosition);
 
-  return length;
+	return length;
 }
 
 void DataBuffer::clear()
 {
-  mStream.clear();
+	mStream.clear();
 }
 
-void DataBuffer::addRaw( const char *b, unsigned size )
+void DataBuffer::addRaw(const char *b, size_t size)
 {
-  mStream.write(b, size);
+	mStream.write(b, size);
 }
 
-bool DataBuffer::readRaw( char *b, unsigned size )
+bool DataBuffer::readRaw(char *b, size_t size)
 {
-  unsigned currentPosition = mStream.getReadPosition();
+	size_t currentPosition = mStream.getReadPosition();
 
-  mStream.setReadPosition(ReadWriteBuffer::End);
-  unsigned length = mStream.getReadPosition();
-  mStream.setReadPosition(currentPosition);
+	mStream.setReadPosition(ReadWriteBuffer::End);
+	size_t length = mStream.getReadPosition();
+	mStream.setReadPosition(currentPosition);
 
-  if (size > (length - currentPosition))
-    return false;
+	if(size > (length - currentPosition))
+		return false;
 
-  mStream.read(b, size);
-  return true;
+	mStream.read(b, size);
+	return true;
 }
 
 template <>
