+++ src/ds/data/font_list.cpp
@@ -8,7 +8,7 @@
 namespace ds {
 
 namespace {
-const int			EMPTY_ID(0);
+const size_t		EMPTY_ID(0);
 const std::string	EMPTY_SZ("");
 }
 
@@ -26,17 +26,16 @@ void FontList::clear(){
 
 void FontList::installFont(const std::string& filePath, const std::string& fontName, const std::string& shortName){
 	mEngine.getPangoFontService().loadFont(filePath, fontName);
-	const int id = mData.size() + 1;
+	const size_t id = mData.size() + 1;
 	mData[id] = Entry(filePath, fontName, shortName);
 }
 
 void FontList::registerFont(const std::string& fontName, const std::string& shortName){
-	const int id = mData.size() + 1;
+	const size_t id = mData.size() + 1;
 	mData[id] = Entry(fontName, fontName, shortName);
 }
 
-int FontList::getIdFromName(const std::string& n) const
-{
+size_t FontList::getIdFromName(const std::string& n) const {
 	if(mData.empty()) return EMPTY_ID;
 	for(auto it = mData.begin(), end = mData.end(); it != end; ++it) {
 		if(it->second.mFilePath == n) return it->first;
@@ -45,7 +44,7 @@ int FontList::getIdFromName(const std::string& n) const
 	return EMPTY_ID;
 }
 
-const std::string& FontList::getFilePathFromId(const int id) const {
+const std::string& FontList::getFilePathFromId(const size_t id) const {
 	if(mData.empty()) return EMPTY_SZ;
 	auto it = mData.find(id);
 	if(it != mData.end()) {
