+++ src/ds/util/exif.cpp
@@ -465,7 +465,7 @@ int easyexif::EXIFInfo::parseFrom(const unsigned char *buf, unsigned len) {
 }
 
 int easyexif::EXIFInfo::parseFrom(const string &data) {
-	return parseFrom((const unsigned char *)data.data(), data.length());
+	return parseFrom((const unsigned char *)data.data(), (unsigned int)data.length());
 }
 
 //
